requirements.txt
Django==4.2.3
gunicorn==20.1.0
djangorestframework==3.14.0
django-rest-knox==4.2.0